move loading/writing of chords
